Convert ModalConfigDatetime to Typescript
