+++ package.json
@@ -10,6 +10,7 @@
     "@types/lodash": "^4.14.176",
     "@types/react": "^17.0.33",
     "@types/react-dom": "^17.0.10",
+    "@types/react-modal": "^3.13.1",
     "@vx/gradient": "0.0.161",
     "@vx/group": "0.0.161",
     "@vx/hierarchy": "0.0.161",
@@ -55,7 +56,7 @@
     "react-lazyload": "^2.2.7",
     "react-leaflet": "^1.9.1",
     "react-leaflet-markercluster": "^1.1.8",
-    "react-modal": "^3.4.5",
+    "react-modal": "^3.14.4",
     "react-pig": "git+https://github.com/LibrePhotos/react-pig.git",
     "react-placeholder": "^3.0.1",
     "react-player": "2.9.0",
